- Updated SLADE Visual Studio project/solution, I'll probably be using this from now on in Windows - Cleaned up/fixed code for compilation in Visual Studio - Fixed the file browser tab under Windows (double clicking a drive in the browser now properly opens it)
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@324 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/ArchiveEntry.cpp
@@ -402,7 +402,7 @@ void ArchiveEntry::detectType(bool data_check, bool force) {
 		parent->detectEntryType(this);
 
 	// Doom flat
-	for (int a = 0; a < n_valid_flat_sizes; a++) {
+	for (size_t a = 0; a < n_valid_flat_sizes; a++) {
 		if (size == valid_flat_size[a][0] * valid_flat_size[a][1])
 			type = ETYPE_FLAT;
 	}
